Update service.sh
added new EAS properties and refactored the way to execute some functions.

+++ Apocalyptic_Tweaks/service.sh
@@ -3,7 +3,7 @@ MODDIR="${0%/*}"
 
 # Aguarda o sistema inicializar
 while [ "$(getprop sys.boot_completed)" != "1" ]; do
-  sleep 10
+  sleep 5
 done
 
 # Executa o script de reset da bateria
@@ -16,20 +16,29 @@ echo 0 > /proc/sys/kernel/sched_schedstats
 echo 64 > /proc/sys/kernel/random/read_wakeup_threshold
 echo "off" > /proc/sys/kernel/printk_devkmsg
 
+# Ajustar prefer_idle para background tasks
+STUNE="/dev/stune"
+if [ -d "$STUNE" ]; then
+    [ -f "$STUNE/background/schedtune.prefer_idle" ] && echo 1 > "$STUNE/background/schedtune.prefer_idle"
+    [ -f "$STUNE/system-background/schedtune.prefer_idle" ] && echo 1 > "$STUNE/system-background/schedtune.prefer_idle"
+fi
+
 # Ativa TTWU_QUEUE no agendador
-echo TTWU_QUEUE > /sys/kernel/debug/sched_features
+[ -e /sys/kernel/debug/sched_features ] && echo TTWU_QUEUE > /sys/kernel/debug/sched_features
 
 sleep 60
 
 # Ajusta swappiness
-echo 100 > /dev/memcg/memory.swappiness
-echo 40 > /dev/memcg/system/memory.swappiness
-echo 50 > /dev/memcg/apps/memory.swappiness
+[ -f /dev/memcg/memory.swappiness ] && echo 100 > /dev/memcg/memory.swappiness
+[ -f /dev/memcg/system/memory.swappiness ] && echo 40 > /dev/memcg/system/memory.swappiness
+[ -f /dev/memcg/apps/memory.swappiness ] && echo 50 > /dev/memcg/apps/memory.swappiness
 
 # Limpa processos do memcg system
-while read -r pid; do
-  echo "$pid" > /dev/memcg/cgroup.procs
-done < /dev/memcg/system/cgroup.procs
+if [ -f /dev/memcg/system/cgroup.procs ] && [ -f /dev/memcg/cgroup.procs ]; then
+  while read -r pid; do
+    echo "$pid" > /dev/memcg/cgroup.procs
+  done < /dev/memcg/system/cgroup.procs
+fi
 
 # Move processos específicos para o memcg system
 for process in system_server surfaceflinger \
@@ -39,8 +48,11 @@ for process in system_server surfaceflinger \
   android.hardware.graphics.composer@2.3-service \
   android.hardware.graphics.composer@2.4-service \
   vendor.qti.hardware.display.composer-service; do
-  pid=$(pidof "$process")
-  [ -n "$pid" ] && echo "$pid" > /dev/memcg/system/cgroup.procs
+
+  pid_list=$(pidof "$process")
+  for pid in $pid_list; do
+    [ -n "$pid" ] && echo "$pid" > /dev/memcg/system/cgroup.procs
+  done
 done
 
 exit 0
